+++ ptai-cli-plugin/src/test/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/cli/BaseIT.java
@@ -22,6 +22,7 @@ import java.util.UUID;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
 
+@Tag("integration")
 class BaseIT {
     /**
      * Temporal folder where three subfolders will be created: for