+++ ptai-jenkins-plugin/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/plugin/jenkins/reports/Json.java
@@ -9,7 +9,6 @@ import org.jenkinsci.Symbol;
 import org.jetbrains.annotations.NotNull;
 import org.kohsuke.stapler.DataBoundConstructor;
 
-@EqualsAndHashCode
 @ToString
 public class Json extends BaseReport {
     @Getter