+++ ptai-jenkins-plugin/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/plugin/jenkins/reports/Report.java
@@ -13,7 +13,6 @@ import org.kohsuke.stapler.DataBoundConstructor;
 
 import java.util.Arrays;
 
-@EqualsAndHashCode
 @ToString
 public class Report extends BaseReport {
     @Getter